+++ src/Client.php
@@ -2,10 +2,12 @@
 
 namespace Spawnia\Sailor;
 
+use stdClass;
+
 interface Client
 {
     /**
      * Execute a GraphQL query against an endpoint and return a Response.
      */
-    public function request(string $query, \stdClass $variables = null): Response;
+    public function request(string $query, stdClass $variables = null): Response;
 }