+++ examples/custom-types/expected/Operations/MyBenSampoEnumQuery.php
@@ -43,6 +43,6 @@ class MyBenSampoEnumQuery extends \Spawnia\Sailor\Operation
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }