+++ examples/custom-types/expected/Operations/MyEnumInputQuery/WithEnumInput/EnumObject.php
@@ -50,6 +50,6 @@ class EnumObject extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../../sailor.php');
     }
 }