+++ examples/custom-types/expected/TypeConverters/CustomDateConverter.php
@@ -36,6 +36,6 @@ class CustomDateConverter implements \Spawnia\Sailor\Convert\TypeConverter
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }