+++ examples/custom-types/expected/TypeConverters/CustomEnumConverter.php
@@ -31,6 +31,6 @@ class CustomEnumConverter implements \Spawnia\Sailor\Convert\TypeConverter
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }