+++ examples/custom-types/expected/Types/DefaultEnum.php
@@ -16,6 +16,6 @@ class DefaultEnum
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }