+++ examples/custom-types/expected/Types/EnumInput.php
@@ -47,6 +47,6 @@ class EnumInput extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }