+++ examples/input/expected/Types/SomeInput.php
@@ -61,6 +61,6 @@ class SomeInput extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }