+++ examples/polymorphic/expected/Operations/AllMembers/AllMembers.php
@@ -45,6 +45,6 @@ class AllMembers extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../sailor.php');
     }
 }