+++ examples/polymorphic/expected/Operations/NodeMembers/Members/Organization.php
@@ -34,6 +34,6 @@ class Organization extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../../sailor.php');
     }
 }