+++ examples/polymorphic/expected/Operations/NodeMembers/NodeMembers.php
@@ -45,6 +45,6 @@ class NodeMembers extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../sailor.php');
     }
 }