+++ examples/polymorphic/expected/Operations/NodeWithFragments/Node/Node/Node/User.php
@@ -50,6 +50,6 @@ class User extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../../../../sailor.php');
     }
 }