+++ examples/polymorphic/expected/Operations/NodeWithFragments/Node/Node/Post.php
@@ -46,6 +46,6 @@ class Post extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../../../sailor.php');
     }
 }