+++ examples/polymorphic/expected/Operations/UserOrPost/Node/Post.php
@@ -50,6 +50,6 @@ class Post extends \Spawnia\Sailor\ObjectLike
 
     public static function config(): string
     {
-        return __DIR__ . '/../../../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../../../sailor.php');
     }
 }