+++ examples/polymorphic/expected/Types/__TypeKind.php
@@ -22,6 +22,6 @@ class __TypeKind
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }