+++ examples/simple/expected/Operations/MyObjectNestedQuery.php
@@ -44,6 +44,6 @@ class MyObjectNestedQuery extends \Spawnia\Sailor\Operation
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }