+++ examples/simple/expected/Operations/MyObjectQuery.php
@@ -41,6 +41,6 @@ class MyObjectQuery extends \Spawnia\Sailor\Operation
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }