+++ examples/simple/expected/Operations/NestedWithFragments.php
@@ -65,6 +65,6 @@ class NestedWithFragments extends \Spawnia\Sailor\Operation
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }