+++ examples/simple/expected/Operations/TwoArgs.php
@@ -48,6 +48,6 @@ class TwoArgs extends \Spawnia\Sailor\Operation
 
     public static function config(): string
     {
-        return __DIR__ . '/../../sailor.php';
+        return \Safe\realpath(__DIR__ . '/../../sailor.php');
     }
 }