+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for types and fields named equivalent to PHP keywords like `type Switch` or `{ print { subfield } }`.
+- Add support for types and fields named equivalent to PHP keywords like `type Switch` or `{ print { subfield } }`
 
 ### Removed